Organize unit tests and add new cases for GetTxTimestamp and ApoptosisResp
